refactor: session_manager, config_manager initialize refactor

+++ lib/myboj/status_manager.rb
@@ -8,22 +8,34 @@ module Myboj
 
     def initialize
       super
+      create_default_file_if_not_exist
+      load_file
+    end
 
+    def create_default_file_if_not_exist
       unless File.exist?(".myboj")
-        File.open(".myboj", "w") do |file|
-          file.puts <<~CONTENT
-            language: python
-            problem: -1
-          CONTENT
-        end
+        File.copy_stream(
+          File.join(Myboj.root_dir, "templates", "default_.myboj"),
+          ".myboj"
+        )
+        true
       end
 
-      File.open(".myboj", "r") do |file|
-        file.readlines.each do |line|
-          key, value = line.chomp.split(": ")
-          self[key] = value
+      false
+    end
+
+    def load_file
+      if File.exist?(".myboj")
+        File.open(".myboj", "r") do |file|
+          file.readlines.each do |line|
+            key, value = line.chomp.split(": ")
+            self[key] = value
+          end
         end
+        true
       end
+
+      false
     end
 
     def persist
